% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_pvec.R
\name{future_pvec}
\alias{future_pvec}
\title{Parallelize a Vector Map Function using Forking}
\usage{
future_pvec(v, FUN, ...)
}
\arguments{
\item{v}{The vector to operate on.}

\item{FUN}{The function to call on each partion of the vector \code{v}.
This function must return the same number of elements as the input,
and it must support one or more elements as input.}

\item{\dots}{Any further arguments passed to \code{FUN()} after each
vector element.}
}
\value{
The value of \code{FUN(v, ...)}.
}
\description{
Parallelize a Vector Map Function using Forking
}
